requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scipy
scikit-learn
wordcloud
nltk
statsmodels
xlsxwriter
chardet
yfinance
alpha_vantage
wikipedia-api
requests>=2.25.1
python-dotenv
